requirements.txt
# RAG-AIOps Framework — Python Dependencies
# Compatible with Python 3.10+

# ── Core Framework ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# ── LLM & Embeddings ───────────────────────────────────────────
openai==1.30.1                  # GPT-4 / text-embedding-ada-002
anthropic==0.26.0               # Claude (alternative LLM)
tiktoken==0.7.0                 # Token counting

# ── Vector Store & Retrieval ───────────────────────────────────
faiss-cpu==1.8.0                # FAISS ANN index (use faiss-gpu for GPU)
sentence-transformers==3.0.1    # Cross-encoder re-ranking (ms-marco-MiniLM)
numpy==1.26.4

# ── Document Processing ────────────────────────────────────────
langchain==0.2.1                # Document loaders & text splitters
langchain-community==0.2.1
langchain-openai==0.1.8
pypdf==4.2.0                    # PDF ingestion
python-docx==1.1.2              # DOCX ingestion
markdown==3.6                   # Markdown ingestion
beautifulsoup4==4.12.3          # HTML cleaning

# ── Data & Storage ─────────────────────────────────────────────
redis==5.0.4                    # Embedding query cache
psycopg2-binary==2.9.9          # PostgreSQL metadata store
sqlalchemy==2.0.30
alembic==1.13.1

# ── Observability & Logging ────────────────────────────────────
structlog==24.1.0
prometheus-client==0.20.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0

# ── NLP & ML Utilities ─────────────────────────────────────────
spacy==3.7.4                    # NER for service/component extraction
langdetect==1.0.9               # Language detection
scikit-learn==1.5.0             # Evaluation metrics

# ── Testing ────────────────────────────────────────────────────
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0                   # FastAPI test client

# ── Dev Tools ──────────────────────────────────────────────────
python-dotenv==1.0.1
PyYAML==6.0.1
click==8.1.7                    # CLI scripts
rich==13.7.1                    # Pretty terminal output
tqdm==4.66.4                    # Progress bars
